throw ConfigException instead of ArgumentError
